Split test files for better parallelization
