--- tests/api/expected-outputs/augsburg_de_pushpage_invalid.json
@@ -1,1 +1,0 @@
-{"status": "denied"}
