--- tests/api/expected-outputs/augsburg_de_pushpage_valid.json
@@ -1,1 +1,0 @@
-{"status": "success"}
