+++ tests/cms/test_duplicate_regions.py
