+++ tests/cms/views/status_code/__init__.py
@@ -1,0 +1,4 @@
+"""
+This package contains simple status code checks for views.
+It is divided into 16 files for better parallelization.
+"""
